feat: Merged all reducer interfaces into a single file for improved organization and maintainability
